feat(menu): make menu responsive at medium screen widths
