+++ package.json
@@ -8,6 +8,7 @@
         "@material-ui/core": "^4.9.14",
         "@material-ui/icons": "^4.9.1",
         "babel-jest": "^26.0.1",
+        "classnames": "^2.2.6",
         "gatsby": "^2.21.22",
         "gatsby-image": "^2.4.5",
         "gatsby-plugin-manifest": "^2.4.2",
